requirements.txt
Flask
Flask-Login
Flask-WTF
mysql-connector-python
python-dotenv
Werkzeug